// frontend/profdiff-web-application/src/app/models/dto/runMetadata.ts
/**
 * Profdiff web-app API
 *
 * Contact: jirkadanek@icloud.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BenchmarkRunMetadata } from './benchmarkRunMetadata';

/**
 * Represents metadata about a profiling report, details about the run BenchmarkRunMetadata, and other statistics extracted from the org.graalvm.profdiff.core.Experiment.
 */
export interface RunMetadata {
  /**
   * Metadata about the profiling run.
   */
  benchmarkRunMetadata: BenchmarkRunMetadata;
  /**
   * Unique identifier for the profiler execution.
   */
  executionId: string;
  /**
   * Total time sampling period of the profiling run.
   */
  totalPeriod?: number;
  /**
   * Sampling period during which GraalVM compiled methods were active.
   */
  graalPeriod?: number;
  /**
   * Total number of GraalVM compilation units in the run.
   */
  compilationUnitsCount: number;
  /**
   * Total number of methods analyzed by profiler.
   */
  proftoolMethodsCount?: number;
}