// frontend/profdiff-web-application/src/app/models/dto/topMethod.ts
/**
 * Profdiff web-app API
 *
 * Contact: jirkadanek@icloud.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Represents a method that is among the top methods by execution time.  It encapsulates subset of relevant data from org.graalvm.profdiff.core.ProftoolMethod. It is used to display a table of top hot methods in UI.  If no profiler information is available, the table will be empty. (not shown at all in UI).
 */
export interface TopMethod {
  /**
   * Percentage of total execution time attributed to this method.
   */
  executionPercentage: string;
  /**
   * Number of cycles (in billions) attributed to this method.
   */
  cycles: string;
  /**
   * Compiler optimization level of the method (0-4).
   */
  level: number;
  /**
   * Unique identifier of the method, if the method is interpreter or stub, then null.
   */
  id?: string;
  /**
   * Full name with arguments types of the method.
   */
  name: string;
}