// frontend/profdiff-web-application/src/app/models/dto/treeResponse.ts
/**
 * Profdiff web-app API
 *
 * Contact: jirkadanek@icloud.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WarningMessage } from './warningMessage';
import { RenderedTreeNode } from './renderedTreeNode';

/**
 * A single tree response representing either an InliningTree, OptimizationTree, or OptimizationContextTree used for the Report and Comparison views.  The tree is converted into a fully structured RenderedTreeNode hierarchy suitable for JSON serialization and interactive rendering on the frontend.
 */
export interface TreeResponse {
  /**
   * The structured representation of the requested compilation tree.
   */
  tree: RenderedTreeNode;
  /**
   * Non-fatal warnings that occurred during experiment parsing.
   */
  warnings?: Array<WarningMessage>;
}