// frontend/profdiff-web-application/src/app/models/dto/methodComparisonPair.ts
/**
 * Profdiff web-app API
 *
 * Contact: jirkadanek@icloud.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { JavaMethod } from './javaMethod';

/**
 * Union of comparative Java methods\' across two different profiling runs.
 */
export interface MethodComparisonPair {
  /**
   * The identifier of the method being compared.
   */
  methodName: string;
  /**
   * The method captured in the first profiling run (nullable).
   */
  methodFromRun1?: JavaMethod;
  /**
   * The method captured in the second profiling run (nullable).
   */
  methodFromRun2?: JavaMethod;
}